refactor!: Moves OnCompleteAction type out of /arc4 namespace
