+++ packages/algo-ts/src/arc4/index.ts
@@ -1,6 +1,7 @@
 import { BaseContract } from '../base-contract'
 import { NoImplementation } from '../internal/errors'
 import { DeliberateAny } from '../internal/typescript-helpers'
+import { OnCompleteActionStr } from '../on-complete-action'
 import { bytes } from '../primitives'
 
 export * from './encoded-types'
@@ -26,40 +27,6 @@ export class Contract extends BaseContract {
  * require: This method CAN ONLY be called when the application is being created
  */
 export type CreateOptions = 'allow' | 'disallow' | 'require'
-/**
- * The possible on complete actions a method can handle, represented as a string
- */
-export type OnCompleteActionStr = 'NoOp' | 'OptIn' | 'ClearState' | 'CloseOut' | 'UpdateApplication' | 'DeleteApplication'
-
-/**
- * The possible on complete actions a method can handle, represented as an integer
- */
-export enum OnCompleteAction {
-  /**
-   * Do nothing after the transaction has completed
-   */
-  NoOp = 0,
-  /**
-   * Opt the calling user into the contract
-   */
-  OptIn = 1,
-  /**
-   * Close the calling user out of the contract
-   */
-  CloseOut = 2,
-  /**
-   * Run the clear state program and forcibly close the user out of the contract
-   */
-  ClearState = 3,
-  /**
-   * Replace the application's approval and clear state programs with the bytes from this transaction
-   */
-  UpdateApplication = 4,
-  /**
-   * Delete the application
-   */
-  DeleteApplication = 5,
-}
 
 /**
  * Type alias for a default argument schema
