+++ packages/algo-ts/src/index.ts
@@ -17,3 +17,4 @@ export { Base64, Ec, Ecdsa, MimcConfigurations, VrfVerify } from './op'
 export { compile, CompiledContract, CompiledLogicSig, CompileContractOptions, CompileLogicSigOptions } from './compiled'
 export { MutableArray } from './mutable-array'
 export { emit } from './arc-28'
+export { OnCompleteAction } from './on-complete-action'
